Menu image and buttons, first SOUND EFFECTS
Uploaded some sound effects (several variations of each).
They are not all good. Feel free to delete bad ones.

For the menu, there are all the buttons for the main menu, one version of menu that is just a complete image with everything and one in which I hid the buttons - I didn't know which was more useful so I just uploaded both